Restructure project, so modules make more sence
