+++ src/tobi.backfrak.de/cmd/samba_exporter/go.mod
@@ -3,17 +3,17 @@ module tobi.backfrak.de/samba_exporter
 require tobi.backfrak.de/internal/commonbl v0.0.0
 replace tobi.backfrak.de/internal/commonbl v0.0.0 => ../../internal/commonbl
 
-require tobi.backfrak.de/cmd/samba_exporter/pipecomunication v0.0.0
-replace tobi.backfrak.de/cmd/samba_exporter/pipecomunication v0.0.0 => ./pipecomunication
+require tobi.backfrak.de/internal/smbexporterbl/pipecomunication v0.0.0
+replace tobi.backfrak.de/internal/smbexporterbl/pipecomunication v0.0.0 => ../../internal/smbexporterbl/pipecomunication
 
-require tobi.backfrak.de/cmd/samba_exporter/statisticsGenerator v0.0.0
-replace tobi.backfrak.de/cmd/samba_exporter/statisticsGenerator v0.0.0 => ./statisticsGenerator
+require tobi.backfrak.de/internal/smbexporterbl/statisticsGenerator v0.0.0
+replace tobi.backfrak.de/internal/smbexporterbl/statisticsGenerator v0.0.0 => ../../internal/smbexporterbl/statisticsGenerator
 
-require tobi.backfrak.de/cmd/samba_exporter/smbstatusreader v0.0.0
-replace tobi.backfrak.de/cmd/samba_exporter/smbstatusreader v0.0.0 => ./smbstatusreader
+require tobi.backfrak.de/internal/smbexporterbl/smbstatusreader v0.0.0
+replace tobi.backfrak.de/internal/smbexporterbl/smbstatusreader v0.0.0 => ../../internal/smbexporterbl/smbstatusreader
 
-require tobi.backfrak.de/cmd/samba_exporter/smbexporter v0.0.0
-replace tobi.backfrak.de/cmd/samba_exporter/smbexporter v0.0.0 => ./smbexporter
+require tobi.backfrak.de/internal/smbexporterbl/smbexporter v0.0.0
+replace tobi.backfrak.de/internal/smbexporterbl/smbexporter v0.0.0 => ../../internal/smbexporterbl/smbexporter
 
 require tobi.backfrak.de/internal/smbstatusout v0.0.0
 replace tobi.backfrak.de/internal/smbstatusout v0.0.0 => ../../internal/smbstatusout
