+++ src/tobi.backfrak.de/cmd/samba_exporter/main.go
@@ -16,10 +16,10 @@ import (
 
 	"github.com/prometheus/client_golang/prometheus"
 	"github.com/prometheus/client_golang/prometheus/promhttp"
-	"tobi.backfrak.de/cmd/samba_exporter/pipecomunication"
-	"tobi.backfrak.de/cmd/samba_exporter/smbexporter"
-	"tobi.backfrak.de/cmd/samba_exporter/smbstatusreader"
-	"tobi.backfrak.de/cmd/samba_exporter/statisticsGenerator"
+	"tobi.backfrak.de/internal/smbexporterbl/pipecomunication"
+	"tobi.backfrak.de/internal/smbexporterbl/smbexporter"
+	"tobi.backfrak.de/internal/smbexporterbl/smbstatusreader"
+	"tobi.backfrak.de/internal/smbexporterbl/statisticsGenerator"
 	"tobi.backfrak.de/internal/commonbl"
 )
