--- src/tobi.backfrak.de/cmd/samba_exporter/pipecomunication/go.mod
@@ -1,10 +1,0 @@
-module tobi.backfrak.de/cmd/samba_exporter/pipecomunication
-
-require tobi.backfrak.de/internal/commonbl v0.0.0
-replace tobi.backfrak.de/internal/commonbl v0.0.0 => ../../../internal/commonbl
-
-require tobi.backfrak.de/cmd/samba_exporter/smbstatusreader v0.0.0
-replace tobi.backfrak.de/cmd/samba_exporter/smbstatusreader v0.0.0 => ../smbstatusreader
-
-require tobi.backfrak.de/internal/smbstatusout v0.0.0
-replace tobi.backfrak.de/internal/smbstatusout v0.0.0 => ../../../internal/smbstatusout
