--- src/tobi.backfrak.de/cmd/samba_exporter/smbstatusreader/go.mod
@@ -1,4 +1,0 @@
-module tobi.backfrak.de/cmd/samba_exporter/smbstatusreader
-
-require tobi.backfrak.de/internal/smbstatusout v0.0.0
-replace tobi.backfrak.de/internal/smbstatusout v0.0.0 => ../../../internal/smbstatusout
