+++ src/tobi.backfrak.de/internal/smbexporterbl/pipecomunication/errors.go
