+++ src/tobi.backfrak.de/internal/smbexporterbl/pipecomunication/errors_test.go
