+++ src/tobi.backfrak.de/internal/smbexporterbl/pipecomunication/go.mod
@@ -1,0 +1,10 @@
+module tobi.backfrak.de/internal/smbexporterbl/pipecomunication
+
+require tobi.backfrak.de/internal/commonbl v0.0.0
+replace tobi.backfrak.de/internal/commonbl v0.0.0 => ../../commonbl
+
+require tobi.backfrak.de/internal/smbexporterbl/smbstatusreader v0.0.0
+replace tobi.backfrak.de/internal/smbexporterbl/smbstatusreader v0.0.0 => ../smbstatusreader
+
+require tobi.backfrak.de/internal/smbstatusout v0.0.0
+replace tobi.backfrak.de/internal/smbstatusout v0.0.0 => ../../smbstatusout
