+++ src/tobi.backfrak.de/internal/smbexporterbl/pipecomunication/handler.go
@@ -10,7 +10,7 @@ import (
 	"sync"
 	"time"
 
-	"tobi.backfrak.de/cmd/samba_exporter/smbstatusreader"
+	"tobi.backfrak.de/internal/smbexporterbl/smbstatusreader"
 	"tobi.backfrak.de/internal/commonbl"
 )
