+++ src/tobi.backfrak.de/internal/smbexporterbl/pipecomunication/handler_test.go
