+++ src/tobi.backfrak.de/internal/smbexporterbl/smbstatusreader/go.mod
@@ -1,0 +1,4 @@
+module tobi.backfrak.de/internal/smbexporterbl/smbstatusreader
+
+require tobi.backfrak.de/internal/smbstatusout v0.0.0
+replace tobi.backfrak.de/internal/smbstatusout v0.0.0 => ../../smbstatusout
