+++ src/tobi.backfrak.de/internal/smbexporterbl/smbstatusreader/reader.go
