+++ src/tobi.backfrak.de/internal/smbexporterbl/statisticsGenerator/go.mod
@@ -1,0 +1,7 @@
+module tobi.backfrak.de/internal/smbexporterbl/statisticsGenerator
+
+require tobi.backfrak.de/internal/smbexporterbl/smbstatusreader v0.0.0
+replace tobi.backfrak.de/internal/smbexporterbl/smbstatusreader v0.0.0 => ../smbstatusreader
+
+require tobi.backfrak.de/internal/smbstatusout v0.0.0
+replace tobi.backfrak.de/internal/smbstatusout v0.0.0 => ../../smbstatusout
